requirements.txt
ansi2html==1.8.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.4.1
Flask==2.2.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.0.3
plotly==5.16.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.2.3
